refactor: migrate quantum cards to modular registry system
- Create new cardRegistry.quantum.ts following the modular pattern used by
  other card categories (misc, ai, platform, etc.)
- Remove inline quantum card declarations from main cardRegistry.ts
- Merge quantum registry into the flat objects via Object.assign before
  registerAllDescriptorCards() — maintains compatibility with current
  cardRegistry.ts structure while using the cleaner modular approach
- All 5 quantum cards remain registered and functional
- TypeScript compilation successful with zero errors

This aligns the quantum cards with upstream's modular architecture in
preparation for PR submission.
